move old matrix to kmk.scanners.digitalio_matrix
