+++ kmk/scanners/__init__.py
@@ -1,3 +1,22 @@
+def intify_coordinate(row, col, len_cols):
+    return len_cols * row + col
+
+
+class DiodeOrientation:
+    '''
+    Orientation of diodes on handwired boards. You can think of:
+    COLUMNS = vertical
+    ROWS = horizontal
+
+    COL2ROW and ROW2COL are equivalent to their meanings in QMK.
+    '''
+
+    COLUMNS = 0
+    ROWS = 1
+    COL2ROW = COLUMNS
+    ROW2COL = ROWS
+
+
 class Scanner:
     '''
     Base class for scanners.
